git-svn-id: svn://svn.sharpdevelop.net/sharpdevelop/trunk@1512 1ccf3a8d-04fe-1044-b7c0-cef0b8235c61

+++ src/Main/Base/Project/Src/Gui/Pads/ProjectBrowser/Commands/DefaultFileNodeCommands.cs
@@ -50,7 +50,7 @@ namespace ICSharpCode.SharpDevelop.Project.Commands
 	
 	public class ExcludeFileFromProject : AbstractMenuCommand
 	{
-		void ExcludeFileNode(FileNode fileNode)
+		public static void ExcludeFileNode(FileNode fileNode)
 		{
 			List<FileNode> dependentNodes = new List<FileNode>();
 			foreach (TreeNode subNode in fileNode.Nodes) {
@@ -60,15 +60,22 @@ namespace ICSharpCode.SharpDevelop.Project.Commands
 			}
 			dependentNodes.ForEach(ExcludeFileNode);
 			
-			ProjectService.RemoveProjectItem(fileNode.Project, fileNode.ProjectItem);
-			fileNode.ProjectItem = null;
-			fileNode.FileNodeStatus = FileNodeStatus.None;
-			if (fileNode.Parent is ExtTreeNode) {
-				((ExtTreeNode)fileNode.Parent).UpdateVisibility();
+			bool isLink = fileNode.IsLink;
+			if (fileNode.ProjectItem != null) { // don't try to exclude same node twice
+				ProjectService.RemoveProjectItem(fileNode.Project, fileNode.ProjectItem);
+			}
+			if (isLink) {
+				fileNode.Remove();
+			} else {
+				fileNode.ProjectItem = null;
+				fileNode.FileNodeStatus = FileNodeStatus.None;
+				if (fileNode.Parent is ExtTreeNode) {
+					((ExtTreeNode)fileNode.Parent).UpdateVisibility();
+				}
 			}
 		}
 		
-		void ExcludeDirectoryNode(DirectoryNode directoryNode)
+		static void ExcludeDirectoryNode(DirectoryNode directoryNode)
 		{
 			if (directoryNode.ProjectItem != null) {
 				ProjectService.RemoveProjectItem(directoryNode.Project, directoryNode.ProjectItem);
